Fixed ReferenceError caused by `to-function` use
`draggable = true` currently causes a `ReferenceError: Invalid left-hand side in assignment` because of the way it's evaluated by `to-function`'s `stringToFunction`. This PR is removing the "magic" but also removing the reliance on `stringToFunction`.

Note: not sure when this occurred, `to-function` has not been updated in awhile. Sortable was working for me a week ago.